requirements.txt
# Core
groq>=0.4.0
tavily-python>=0.5.0
streamlit>=1.39.0

# LangGraph
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# Vector DB
chromadb>=0.5.0
sentence-transformers>=2.2.0

# Evaluation
ragas>=0.1.0
deepeval>=1.0.0

# Export
fpdf2>=2.7.0
markdown>=3.5.0

# API
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.9

# Utilities
pydantic>=2.9.0
python-dotenv>=1.0.0
httpx>=0.27.0
aiofiles>=24.1.0
